backend/pyproject.toml
[project]
name = "stock-analysis-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "langchain[openai]>=1.2.13",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
    "yfinance>=1.2.0",
]
